Move MiniFoundation back to legacy
